prevent delete transaction item if it's posted

+++ app/Http/Controllers/TransactionController.php
@@ -227,6 +227,12 @@ class TransactionController extends Controller
         if(Gate::denies('تعديل بند في المعاملة')) {
             return redirect()->back()->with('error', 'ليس لديك صلاحية تعديل بند في المعاملة');
         }
+        if($item->transaction->containers->first()->invoices->where('type', 'تخليص')->first()) {
+            return redirect()->back()->with('error', 'لا يمكن تعديل بند من معاملة تم إصدار فاتورة لها');
+        }
+        if($item->is_posted) {
+            return redirect()->back()->with('error', 'لا يمكن تعديل بند تم ترحيله مسبقاً');
+        }
 
         $old = $item->toArray();
         $validated = $request->validated();
@@ -284,6 +290,13 @@ class TransactionController extends Controller
         if(Gate::denies('حذف بند من المعاملة')) {
             return redirect()->back()->with('error', 'ليس لديك صلاحية حذف بند من المعاملة');
         }
+        if($item->transaction->containers->first()->invoices->where('type', 'تخليص')->first()) {
+            return redirect()->back()->with('error', 'لا يمكن حذف بند من معاملة تم إصدار فاتورة لها');
+        }
+        if($item->is_posted) {
+            return redirect()->back()->with('error', 'لا يمكن حذف بند تم ترحيله مسبقاً');
+        }
+
         $old = $item->toArray();
         $item->delete();
         logActivity('حذف بند من المعاملة', "تم حذف بند من المعاملة رقم " . $item->transaction->code, $old, null);
